+++ dailyVerse/DVExtension/UIColor+DV.swift
@@ -1,0 +1,59 @@
+//
+//  UIColor+DV.swift
+//  dailyVerse
+//
+//  Created by itamaker on 2017/11/14.
+//  Copyright © 2017年 庫倪. All rights reserved.
+//
+
+import UIKit
+
+/*
+ App 顏色板
+ */
+extension UIColor {
+    
+    @nonobjc class var dlyCoralPink: UIColor {
+        return UIColor(named: "coralPink")!
+    }
+    
+    @nonobjc class var dlySoftBlue: UIColor {
+        return UIColor(named: "softBlue")!
+    }
+    
+    @nonobjc class var dlyBlack12: UIColor {
+        return UIColor(named: "black12")!
+    }
+    
+    @nonobjc class var dlyGunmetal: UIColor {
+        return UIColor(named: "gunmetal")!
+    }
+    
+    @nonobjc class var dlyGreyishBrown: UIColor {
+        return UIColor(named: "greyishBrown")!
+    }
+    
+    @nonobjc class var dlyDark: UIColor {
+        return UIColor(named: "dark")!
+    }
+    
+    @nonobjc class var dlyBlack24: UIColor {
+        return UIColor(named: "black24")!
+    }
+    
+    @nonobjc class var dlyWhite50: UIColor {
+        return UIColor(named: "white50")!
+    }
+    
+    @nonobjc class var dlyWhite: UIColor {
+        return UIColor(named: "white")!
+    }
+    
+    @nonobjc class var dlyWhite0: UIColor {
+        return UIColor(named: "white0")!
+    }
+    
+    @nonobjc class var dlyPaleGrey: UIColor {
+        return UIColor(named: "paleGrey")!
+    }
+}
