+++ dailyVerse/ViewController.swift
@@ -567,131 +567,11 @@ class ViewController: UIViewController, UITableViewDataSource, UITabBarDelegate 
         
         return traditional[longName]!
     }
-
-}
-
-
-
-
-
-/*
- App 顏色板
- */
-
-extension UIColor {
-
-    @nonobjc class var dlyCoralPink: UIColor {
-        return UIColor(named: "coralPink")!
-    }
-
-    @nonobjc class var dlySoftBlue: UIColor {
-        return UIColor(named: "softBlue")!
-    }
-
-    @nonobjc class var dlyBlack12: UIColor {
-        return UIColor(named: "black12")!
-    }
-
-    @nonobjc class var dlyGunmetal: UIColor {
-        return UIColor(named: "gunmetal")!
-    }
-
-    @nonobjc class var dlyGreyishBrown: UIColor {
-        return UIColor(named: "greyishBrown")!
-    }
-
-    @nonobjc class var dlyDark: UIColor {
-        return UIColor(named: "dark")!
-    }
-
-    @nonobjc class var dlyBlack24: UIColor {
-        return UIColor(named: "black24")!
-    }
-
-    @nonobjc class var dlyWhite50: UIColor {
-        return UIColor(named: "white50")!
-    }
-
-    @nonobjc class var dlyWhite: UIColor {
-        return UIColor(named: "white")!
-    }
-
-    @nonobjc class var dlyWhite0: UIColor {
-        return UIColor(named: "white0")!
-    }
-
-    @nonobjc class var dlyPaleGrey: UIColor {
-        return UIColor(named: "paleGrey")!
-    }
-}
-
-/*
- 文字樣式
- */
-extension UIFont {
-
-    @nonobjc class var dlyMainTextStyle: UIFont {
-        return UIFont(name: "PingFangTC-Light", size: 16.0)!
-    }
-
-    @nonobjc class var dlyTodayTextStyle: UIFont {
-        return UIFont(name: "PingFangTC-Semibold", size: 10.0)!
-    }
-
-    @nonobjc class var dlyDateTextStyle: UIFont {
-        return UIFont(name: "PingFangTC-Regular", size: 10.0)!
-    }
-}
-
-
-/*
- setLineSpacing 設定行高 / lineHeightMultiple / 字距
- */
-extension UILabel {
-
-    func typesetting(lineSpacing: CGFloat = 0.0, lineHeightMultiple: CGFloat = 0.0, characterSpacing: CGFloat = 0.0) {
-
-        guard let labelText = self.text else { return }
-
-        let paragraphStyle = NSMutableParagraphStyle()
-        paragraphStyle.lineSpacing = lineSpacing
-        paragraphStyle.lineHeightMultiple = lineHeightMultiple
-
-        let attributedString: NSMutableAttributedString
-        if let labelattributedText = self.attributedText {
-            attributedString = NSMutableAttributedString(attributedString: labelattributedText)
-        } else {
-            attributedString = NSMutableAttributedString(string: labelText)
-        }
-
-        // 處理字距
-        attributedString.addAttribute(NSAttributedStringKey.kern, value: characterSpacing, range: NSRange(location: 0, length: attributedString.length - 1))
-
-        // 處理行高
-        attributedString.addAttribute(NSAttributedStringKey.paragraphStyle, value: paragraphStyle, range: NSMakeRange(0, attributedString.length))
-
-        self.attributedText = attributedString
-    }
-}
-
-/*
- UI Text View 垂直居中
- */
-
-extension UITextView {
-    func centerVertically() {
-        let fittingSize = CGSize(width: bounds.width, height: CGFloat.greatestFiniteMagnitude)
-        let size = sizeThatFits(fittingSize)
-        let topOffset = (bounds.size.height - size.height * zoomScale) / 2
-        let positiveTopOffset = max(1, topOffset)
-        contentOffset.y = -positiveTopOffset
-    }
 }
 
 /*
  IB 追加陰影\圓角等樣式
  */
-
 @IBDesignable
 class DesignableView: UIView {
 }
@@ -704,89 +584,4 @@ class DesignableButton: UIButton {
 class DesignableLabel: UILabel {
 }
 
-extension UIView {
 
-    @IBInspectable
-    var cornerRadius: CGFloat {
-        get {
-            return layer.cornerRadius
-        }
-        set {
-            layer.cornerRadius = newValue
-        }
-    }
-
-    @IBInspectable
-    var borderWidth: CGFloat {
-        get {
-            return layer.borderWidth
-        }
-        set {
-            layer.borderWidth = newValue
-        }
-    }
-
-    @IBInspectable
-    var borderColor: UIColor? {
-        get {
-            if let color = layer.borderColor {
-                return UIColor(cgColor: color)
-            }
-            return nil
-        }
-        set {
-            if let color = newValue {
-                layer.borderColor = color.cgColor
-            } else {
-                layer.borderColor = nil
-            }
-        }
-    }
-
-    @IBInspectable
-    var shadowRadius: CGFloat {
-        get {
-            return layer.shadowRadius
-        }
-        set {
-            layer.shadowRadius = newValue
-        }
-    }
-
-    @IBInspectable
-    var shadowOpacity: Float {
-        get {
-            return layer.shadowOpacity
-        }
-        set {
-            layer.shadowOpacity = newValue
-        }
-    }
-
-    @IBInspectable
-    var shadowOffset: CGSize {
-        get {
-            return layer.shadowOffset
-        }
-        set {
-            layer.shadowOffset = newValue
-        }
-    }
-
-    @IBInspectable
-    var shadowColor: UIColor? {
-        get {
-            if let color = layer.shadowColor {
-                return UIColor(cgColor: color)
-            }
-            return nil
-        }
-        set {
-            if let color = newValue {
-                layer.shadowColor = color.cgColor
-            } else {
-                layer.shadowColor = nil
-            }
-        }
-    }
-}
